Split `highfive/boost.hpp`.
Allowing more precise control over which parts of
boost need to be available.
